+++ luxonis_train/attached_modules/metrics/__init__.py
@@ -1,9 +1,9 @@
 from .base_metric import BaseMetric
 from .confusion_matrix import ConfusionMatrix
+from .embedding_metrics import ClosestIsPositiveAccuracy, MedianDistances
 from .mean_average_precision import MeanAveragePrecision
 from .mean_average_precision_keypoints import MeanAveragePrecisionKeypoints
 from .object_keypoint_similarity import ObjectKeypointSimilarity
-from .pml_metrics import ClosestIsPositiveAccuracy, MedianDistances
 from .torchmetrics import Accuracy, F1Score, JaccardIndex, Precision, Recall
 
 __all__ = [
